feat: add ABC Jobs.postman_collection.json
Add ABC Jobs.postman_collection.json file to the project. This file contains a Postman collection with a request for performance testing. The request is named "Motor de Emparejamiento" and it is a GET request to the "localhost/candidatos" endpoint. This collection is useful for testing the performance of the application.